# Move `Reactive.Banana.Test.Performance to a proper `benchmark` suite