--- a/reactive-banana/reactive-banana.cabal
+++ b/reactive-banana/reactive-banana.cabal
@@ -94,6 +94,7 @@ Test-Suite tests
 
 
 Benchmark benchmark
+  default-language:   Haskell2010
   type:          exitcode-stdio-1.0
   build-depends: base, tasty-bench, reactive-banana, containers, random, tasty
   hs-source-dirs: benchmark