… reactive-banana/src/Reactive/Banana/Test/Performance.hs