I'd like to start doing some performance investigations into `reactive-banana`, but the to do that, I need benchmarks! We do currently have a very basic benchmark in `Reactive.Banana.Test.Performance`, but there's no convenient way to actually run it. I've moved this benchmark out of `src` and into a new top-level `benchmark` folder, and added a `benchmark` component to `reactive-banana.cabal`. This new benchmark component uses the lightweight benchmarking library `tasty-bench` to run the benchmarks. A nice property of this library is that we could also actually use these benchmarks as regression tests, to make sure performance only gets better. My goal of this PR is mostly a refactoring though, so I'll leave that other work.

Here's what running this benchmark looks like:

```
❯ cabal run benchmark
Up to date
All
  netsize = 1
    duration = 1:   OK (0.24s)
      6.66 μs ± 479 ns
    duration = 2:   OK (0.12s)
      13.4 μs ± 1.3 μs
    duration = 4:   OK (0.22s)
      26.7 μs ± 1.8 μs
    duration = 8:   OK (0.22s)
      53.6 μs ± 3.0 μs
    duration = 16:  OK (0.22s)
      106  μs ± 5.7 μs
    duration = 32:  OK (0.11s)
      213  μs ±  21 μs
    duration = 64:  OK (0.22s)
      425  μs ±  30 μs
    duration = 128: OK (0.11s)
      858  μs ±  84 μs
  netsize = 2
    duration = 1:   OK (0.12s)
      7.00 μs ± 697 ns
    duration = 2:   OK (0.12s)
      13.9 μs ± 1.3 μs
    duration = 4:   OK (0.12s)
      27.7 μs ± 2.7 μs
    duration = 8:   OK (0.23s)
      55.1 μs ± 3.3 μs
    duration = 16:  OK (0.11s)
      110  μs ±  11 μs
    duration = 32:  OK (0.22s)
      220  μs ±  15 μs
    duration = 64:  OK (0.12s)
      442  μs ±  44 μs
    duration = 128: OK (0.23s)
      885  μs ±  43 μs
  netsize = 4
    duration = 1:   OK (0.24s)
      7.07 μs ± 351 ns
    duration = 2:   OK (0.12s)
      14.1 μs ± 1.4 μs
    duration = 4:   OK (0.23s)
      27.7 μs ± 1.4 μs
    duration = 8:   OK (0.12s)
      55.4 μs ± 5.4 μs
    duration = 16:  OK (0.23s)
      111  μs ± 5.3 μs
    duration = 32:  OK (0.23s)
      222  μs ±  11 μs
    duration = 64:  OK (0.23s)
      445  μs ±  29 μs
    duration = 128: OK (0.23s)
      892  μs ±  46 μs
  netsize = 8
    duration = 1:   OK (0.13s)
      7.28 μs ± 685 ns
    duration = 2:   OK (0.12s)
      14.3 μs ± 1.3 μs
    duration = 4:   OK (0.24s)
      28.4 μs ± 1.3 μs
    duration = 8:   OK (0.24s)
      56.7 μs ± 3.1 μs
    duration = 16:  OK (0.24s)
      113  μs ± 7.0 μs
    duration = 32:  OK (0.22s)
      228  μs ±  11 μs
    duration = 64:  OK (0.12s)
      453  μs ±  43 μs
    duration = 128: OK (0.12s)
      901  μs ±  87 μs
  netsize = 16
    duration = 1:   OK (0.13s)
      7.54 μs ± 707 ns
    duration = 2:   OK (0.13s)
      14.7 μs ± 1.3 μs
    duration = 4:   OK (0.25s)
      29.2 μs ± 1.8 μs
    duration = 8:   OK (0.12s)
      57.9 μs ± 5.3 μs
    duration = 16:  OK (0.24s)
      115  μs ± 6.3 μs
    duration = 32:  OK (0.12s)
      229  μs ±  22 μs
    duration = 64:  OK (0.12s)
      461  μs ±  43 μs
    duration = 128: OK (0.24s)
      918  μs ±  53 μs
  netsize = 32
    duration = 1:   OK (0.14s)
      8.11 μs ± 707 ns
    duration = 2:   OK (0.13s)
      15.4 μs ± 1.3 μs
    duration = 4:   OK (0.13s)
      29.8 μs ± 2.7 μs
    duration = 8:   OK (0.13s)
      58.8 μs ± 5.4 μs
    duration = 16:  OK (0.11s)
      117  μs ±  11 μs
    duration = 32:  OK (0.24s)
      233  μs ±  13 μs
    duration = 64:  OK (0.24s)
      467  μs ±  23 μs
    duration = 128: OK (0.12s)
      927  μs ±  87 μs
  netsize = 64
    duration = 1:   OK (0.16s)
      9.37 μs ± 665 ns
    duration = 2:   OK (0.13s)
      16.9 μs ± 1.3 μs
    duration = 4:   OK (0.13s)
      31.7 μs ± 3.1 μs
    duration = 8:   OK (0.13s)
      61.6 μs ± 5.3 μs
    duration = 16:  OK (0.24s)
      121  μs ± 9.3 μs
    duration = 32:  OK (0.13s)
      241  μs ±  22 μs
    duration = 64:  OK (0.13s)
      478  μs ±  45 μs
    duration = 128: OK (0.13s)
      952  μs ±  85 μs
  netsize = 128
    duration = 1:   OK (0.30s)
      34.3 μs ± 3.4 μs
    duration = 2:   OK (0.15s)
      64.0 μs ± 5.5 μs
    duration = 4:   OK (0.14s)
      122  μs ±  11 μs
    duration = 8:   OK (0.26s)
      241  μs ±  12 μs
    duration = 16:  OK (0.14s)
      505  μs ±  44 μs
    duration = 32:  OK (0.14s)
      1.00 ms ±  85 μs
    duration = 64:  OK (0.13s)
      1.99 ms ± 174 μs
    duration = 128: OK (0.13s)
      4.00 ms ± 336 μs

All 64 tests passed (11.10s)
```