@mitchellwrosen 2021-12-18
Nice! LGTM, just looks like a missing default-language.

Btw, for more throwaway benchmarking, you may be interested in this package I just uploaded to hackage:

https://hackage.haskell.org/package/park-bench