@ocharles 2021-12-19
> Btw, for more throwaway benchmarking, you may be interested in this package I just uploaded to hackage:

Very cool! Have you posted this on the Haskell Discourse forum? Would be great to announce it - the more tools for performance, the better!