@luke-clifton 2021-12-19
Slightly off topic...

> https://hackage.haskell.org/package/park-bench

Seems the github repo is private (or incorrect in the cabal file?). I'm getting a 404.